+++ backend/openedx_ai_extensions/tasks.py
@@ -2,4 +2,5 @@
 Celery tasks that will be registered.
 """
 # pylint: disable=unused-import
-from openedx_ai_extensions.workflows.orchestrators import _execute_orchestrator_async  # noqa: F401
+from openedx_ai_extensions.workflows.orchestrators.session_based_orchestrator import \
+    _execute_orchestrator_async  # noqa: F401
