+++ backend/openedx_ai_extensions/workflows/orchestrators/__init__.py
@@ -1,0 +1,10 @@
+"""
+Orchestrators package for AI workflow execution.
+
+This package provides the base orchestrator and concrete implementations.
+"""
+from .base_orchestrator import BaseOrchestrator
+
+__all__ = [
+    "BaseOrchestrator",
+]
