+++ backend/openedx_ai_extensions/workflows/orchestrators/orchestrators.py
@@ -4,20 +4,13 @@ Base classes to hold the logic of execution in ai workflows
 """
 import json
 import logging
-import sys
 import time
-from typing import TYPE_CHECKING
-
-from celery import shared_task
-from celery.exceptions import SoftTimeLimitExceeded
-from eventtracking import tracker
 
 from openedx_ai_extensions.processors import (
     ContentLibraryProcessor,
     EducatorAssistantProcessor,
     LLMProcessor,
     OpenEdXProcessor,
-    SubmissionProcessor,
 )
 from openedx_ai_extensions.utils import is_generator
 from openedx_ai_extensions.xapi.constants import (
@@ -26,170 +19,10 @@ from openedx_ai_extensions.xapi.constants import (
     EVENT_NAME_WORKFLOW_INTERACTED,
 )
 
-if TYPE_CHECKING:
-    from openedx_ai_extensions.workflows.models import AIWorkflow, AIWorkflowSession
+from .base_orchestrator import BaseOrchestrator
+from .session_based_orchestrator import SessionBasedOrchestrator
 
 logger = logging.getLogger(__name__)
-
-
-@shared_task(
-    name="openedx_ai_extensions.workflows.execute_orchestrator",
-    bind=True,
-    time_limit=300,
-    soft_time_limit=270
-)
-def _execute_orchestrator_async(task_self, session_id, action, params=None):
-    """
-    Execute an orchestrator action asynchronously.
-
-    Args:
-        task_self: Celery task instance (bound)
-        session_id: UUID of the AIWorkflowSession
-        action: Method name to call on the orchestrator (e.g., 'run')
-        params: Dictionary of parameters to pass to the action method
-
-    Returns:
-        Result from the orchestrator action method
-    """
-    from openedx_ai_extensions.workflows.models import AIWorkflowSession  # pylint: disable=import-outside-toplevel
-
-    task_id = task_self.request.id
-    params = params or {}
-
-    try:
-        # 1. Get the session from the database
-        session = AIWorkflowSession.objects.select_related('scope', 'profile', 'user').get(id=session_id)
-
-        # 2. Build context from session
-        context = {
-            'course_id': str(session.course_id),
-            'location_id': str(session.location_id),
-        }
-
-        # 3. Resolve and instantiate orchestrator via centralized factory
-        orchestrator_name = session.profile.orchestrator_class
-        try:
-            orchestrator = BaseOrchestrator.get_orchestrator(
-                workflow=session.scope,
-                user=session.user,
-                context=context,
-            )
-        except (AttributeError, TypeError) as exc:
-            logger.error(
-                f"Task {task_id}: Failed to resolve orchestrator: {exc}",
-                exc_info=True,
-            )
-            raise
-
-        # 4. Validate action exists
-        if not hasattr(orchestrator, action):
-            error_msg = f"Orchestrator '{orchestrator_name}' does not have method '{action}'"
-            logger.error(f"Task {task_id}: {error_msg}")
-            raise AttributeError(error_msg)
-
-        # 5. Call the action method with params
-        orchestrator_method = getattr(orchestrator, action)
-        logger.info(f"Task {task_id}: Executing {orchestrator_name}.{action} for session {session_id}")
-        result = orchestrator_method(**params)
-
-        # 6. Update session metadata with result
-        session.metadata['task_result'] = result
-        session.metadata['task_status'] = 'completed'
-        session.save(update_fields=['metadata'])
-
-        logger.info(f"Task {task_id}: Completed successfully")
-        return result
-
-    except SoftTimeLimitExceeded:
-        logger.error(f"Task {task_id}: Soft time limit exceeded for session {session_id}")
-        session.metadata['task_status'] = 'timeout'
-        session.metadata['task_error'] = 'Task exceeded time limit'
-        session.save(update_fields=['metadata'])
-        raise
-
-    except AIWorkflowSession.DoesNotExist:
-        logger.error(f"Task {task_id}: Session {session_id} not found")
-        raise
-
-    except Exception as e:
-        logger.error(f"Task {task_id}: Error executing {action} for session {session_id}: {str(e)}")
-        session.metadata['task_status'] = 'error'
-        session.metadata['task_error'] = str(e)
-        session.save(update_fields=['metadata'])
-        raise
-
-
-class BaseOrchestrator:
-    """Base class for workflow orchestrators."""
-
-    def __init__(self, workflow, user, context):
-        self.workflow = workflow
-        self.user = user
-        self.profile = workflow.profile
-        self.location_id = context.get("location_id", None)
-        self.course_id = context.get("course_id", None)
-
-    def _emit_workflow_event(self, event_name: str) -> None:
-        """
-        Emit an xAPI event for this workflow.
-
-        Args:
-            event_name: The event name constant (e.g., EVENT_NAME_WORKFLOW_COMPLETED)
-        """
-
-        tracker.emit(event_name, {
-            "workflow_id": str(self.workflow.id),
-            "action": self.workflow.action,
-            "course_id": str(self.course_id),
-            "profile_name": self.profile.slug,
-            "location_id": str(self.location_id),
-        })
-
-    def run(self, input_data):
-        raise NotImplementedError("Subclasses must implement run method")
-
-    @classmethod
-    def get_orchestrator(cls, *, workflow, user, context):
-        """
-        Resolve and instantiate an orchestrator for the given workflow.
-
-        This factory method centralizes orchestrator lookup and validation.
-        It ensures that the resolved class exists and is a subclass of
-        BaseOrchestrator, providing a single, consistent entry point
-        for orchestrator creation across the codebase.
-
-        Args:
-            workflow: AIWorkflowScope instance that defines the workflow configuration.
-            user: User for whom the workflow is being executed.
-            context: Dictionary with runtime context (e.g. course_id, location_id).
-
-        Returns:
-            BaseOrchestrator: An instantiated orchestrator for the given workflow.
-
-        Raises:
-            AttributeError: If the configured orchestrator class cannot be found.
-            TypeError: If the resolved class is not a subclass of BaseOrchestrator.
-        """
-        orchestrator_name = workflow.profile.orchestrator_class
-
-        try:
-            module = sys.modules[__name__]
-            orchestrator_class = getattr(module, orchestrator_name)
-        except AttributeError as exc:
-            raise AttributeError(
-                f"Orchestrator class '{orchestrator_name}' not found"
-            ) from exc
-
-        if not issubclass(orchestrator_class, BaseOrchestrator):
-            raise TypeError(
-                f"{orchestrator_name} is not a subclass of BaseOrchestrator"
-            )
-
-        return orchestrator_class(
-            workflow=workflow,
-            user=user,
-            context=context,
-        )
 
 
 class MockResponse(BaseOrchestrator):
@@ -300,36 +133,6 @@ class DirectLLMResponse(BaseOrchestrator):
         return response_data
 
 
-class SessionBasedOrchestrator(BaseOrchestrator):
-    """Orchestrator that provides session-based LLM responses."""
-
-    def __init__(self, workflow, user, context):
-        from openedx_ai_extensions.workflows.models import AIWorkflowSession  # pylint: disable=import-outside-toplevel
-
-        super().__init__(workflow, user, context)
-        self.session, _ = AIWorkflowSession.objects.get_or_create(
-            user=self.user,
-            scope=self.workflow,
-            profile=self.workflow.profile,
-            defaults={},
-        )
-
-    def clear_session(self, _):
-        self.session.delete()
-        return {
-            "response": "",
-            "status": "session_cleared",
-        }
-
-    def _get_submission_processor(self):
-        return SubmissionProcessor(
-            self.profile.processor_config, self.session
-        )
-
-    def run(self, input_data):
-        raise NotImplementedError("Subclasses must implement run method")
-
-
 class EducatorAssistantOrchestrator(SessionBasedOrchestrator):
     """
     Orchestrator for educator assistant workflows.
@@ -401,32 +204,6 @@ class EducatorAssistantOrchestrator(SessionBasedOrchestrator):
                 'tokens_used': llm_result.get('tokens_used'),
                 'model_used': llm_result.get('model_used')
             }
-        }
-
-    def run_async(self, input_data):
-        """
-        Launch async task to execute the run method.
-
-        Args:
-            input_data: Input data to pass to the run method
-        """
-
-        self.session.course_id = self.course_id
-        self.session.location_id = self.location_id
-        self.session.save()
-
-        task = _execute_orchestrator_async.delay(
-            session_id=self.session.id,
-            action='run',
-            params={
-                "input_data": input_data,
-            }
-        )
-
-        return {
-            'status': 'processing',
-            'task_id': task.id,
-            'message': 'AI workflow has started'
         }
 
     def get_run_status(self, input_data):  # pylint: disable=unused-argument
